@@ -1,6 +1,6 @@
 {
   "name": "glob-parent",
-  "version": "5.1.2",
+  "version": "6.0.2",
   "description": "Extract the non-magic parent path from a glob string.",
   "author": "Gulp Team <team@gulpjs.com> (https://gulpjs.com/)",
   "contributors": [
@@ -10,7 +10,7 @@
   "repository": "gulpjs/glob-parent",
   "license": "ISC",
   "engines": {
-    "node": ">= 6"
+    "node": ">=10.13.0"
   },
   "main": "index.js",
   "files": [
@@ -20,20 +20,26 @@
   "scripts": {
     "lint": "eslint .",
     "pretest": "npm run lint",
-    "test": "nyc mocha --async-only",
-    "azure-pipelines": "nyc mocha --async-only --reporter xunit -O output=test.xunit",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls"
+    "test": "nyc mocha --async-only"
   },
   "dependencies": {
-    "is-glob": "^4.0.1"
+    "is-glob": "^4.0.3"
   },
   "devDependencies": {
-    "coveralls": "^3.0.11",
-    "eslint": "^2.13.1",
-    "eslint-config-gulp": "^3.0.1",
-    "expect": "^1.20.2",
-    "mocha": "^6.0.2",
-    "nyc": "^13.3.0"
+    "eslint": "^7.0.0",
+    "eslint-config-gulp": "^5.0.0",
+    "expect": "^26.0.1",
+    "mocha": "^7.1.2",
+    "nyc": "^15.0.1"
+  },
+  "nyc": {
+    "reporter": [
+      "lcov",
+      "text-summary"
+    ]
+  },
+  "prettier": {
+    "singleQuote": true
   },
   "keywords": [
     "glob",
