@@ -1,7 +1,7 @@
 {
   "name": "picomatch",
   "description": "Blazing fast and accurate glob matcher written in JavaScript, with no dependencies and full support for standard and extended Bash glob features, including braces, extglobs, POSIX brackets, and regular expressions.",
-  "version": "2.3.1",
+  "version": "4.0.3",
   "homepage": "https://github.com/micromatch/picomatch",
   "author": "Jon Schlinkert (https://github.com/jonschlinkert)",
   "funding": "https://github.com/sponsors/jonschlinkert",
@@ -12,11 +12,13 @@
   "license": "MIT",
   "files": [
     "index.js",
+    "posix.js",
     "lib"
   ],
+  "sideEffects": false,
   "main": "index.js",
   "engines": {
-    "node": ">=8.6"
+    "node": ">=12"
   },
   "scripts": {
     "lint": "eslint --cache --cache-location node_modules/.cache/.eslintcache --report-unused-disable-directives --ignore-path .gitignore .",
@@ -26,11 +28,11 @@
     "test:cover": "nyc npm run mocha"
   },
   "devDependencies": {
-    "eslint": "^6.8.0",
+    "eslint": "^8.57.0",
     "fill-range": "^7.0.1",
     "gulp-format-md": "^2.0.0",
-    "mocha": "^6.2.2",
-    "nyc": "^15.0.0",
+    "mocha": "^10.4.0",
+    "nyc": "^15.1.0",
     "time-require": "github:jonschlinkert/time-require"
   },
   "keywords": [
